+++ R/PriceIndices.R
@@ -8,6 +8,8 @@
 #' @docType package
 #' @name PriceIndices
 #' @section Data sets in the package and generating artificial scanner data sets:
+#' \code{\link{dataAGGR}}
+#'
 #' \code{\link{dataMATCH}}
 #'
 #' \code{\link{dataCOICOP}}
@@ -28,6 +30,8 @@
 #' \code{\link{data_check}} 
 #'
 #' \code{\link{data_preparing}} 
+#'
+#' \code{\link{data_aggregating}} 
 #'
 #' \code{\link{data_unit}} 
 #'