+++ R/dataMATCH.R
@@ -1,4 +1,4 @@
-#' Artificial data set on product sales
+#' An artificial scanner data set for product matching
 #'
 #' A collection of scanner data on the sale of sample artificial products.  
 #' @format A data frame with 7 columns and 30 rows. The used variables are as follows: